Updated @DELETE for Submission and Review controllers so users can only delete their own submissions and reviews.

+++ ThirdPlace-Backend/src/main/java/com/CherrySystems/ThirdPlace_Backend/controllers/SubmissionController.java
@@ -1,6 +1,7 @@
 package com.CherrySystems.ThirdPlace_Backend.controllers;
 
 import com.CherrySystems.ThirdPlace_Backend.models.Category;
+import com.CherrySystems.ThirdPlace_Backend.models.Review;
 import com.CherrySystems.ThirdPlace_Backend.models.Submission;
 import com.CherrySystems.ThirdPlace_Backend.models.User;
 import com.CherrySystems.ThirdPlace_Backend.models.dto.SubmissionFormDTO;
@@ -130,7 +131,22 @@ public class SubmissionController {
 
     //Deletes submissions in Submission Repository by finding the submission by its ID#
     @DeleteMapping("/{id}")
-    public void deleteSubmission(@PathVariable Integer id) {
-        submissionRepository.deleteById(id);
+    public ResponseEntity<?> deleteSubmission(@PathVariable Integer id, HttpSession session) {
+
+        User user = authenticationController.getUserFromSession(session);
+        if (user == null) {
+            return ResponseEntity.badRequest().body("User is not logged in.");
+        }
+
+        //Find submission by ID in repository
+        Submission submissionById = submissionRepository.findById(id).get();
+
+        if (!(submissionById.getUser()).equals(user)) {
+            return ResponseEntity.badRequest().body("Submission can only be deleted by posting user.");
+        } else {
+            submissionRepository.deleteById(id);
+        }
+        return  ResponseEntity.ok("Submission deleted.");
     }
+
 }
